Cygwin: dll_init: use SIZE_T instead of DWORD for size
The RegionSize member of the MEMORY_BASIC_INFORMATION struct is of type
SIZE_T, and it may be larger than will fit in a DWORD (I observed
0x200000000).  This resulted in an error due to trying to reserve 0
bytes from VirtualAllloc.

Fixes: 8d777a13fcf4 ("* dll_init.cc (reserve_at, release_at): New functions.")
Addresses: https://cygwin.com/pipermail/cygwin/2025-May/258154.html
Reported-by: Yuyi Wang <Strawberry_Str@hotmail.com>

+++ winsup/cygwin/dll_init.cc
@@ -633,7 +633,7 @@ dll_list::track_self ()
 static PVOID
 reserve_at (PCWCHAR name, PVOID here, PVOID dll_base, DWORD dll_size)
 {
-  DWORD size;
+  SIZE_T size;
   MEMORY_BASIC_INFORMATION mb;
 
   if (!VirtualQuery (here, &mb, sizeof (mb)))
